refactor(shadcn): better api
