+++ packages/shadcn/src/registry/api.ts
@@ -2,24 +2,14 @@ import { createHash } from "crypto"
 import { promises as fs } from "fs"
 import { homedir } from "os"
 import path from "path"
-import {
-  buildUrlAndHeadersForRegistryItem,
-  resolveRegistryUrl,
-} from "@/src/registry/builder"
 import { clearRegistryContext } from "@/src/registry/context"
 import { parseRegistryAndItemFromString } from "@/src/registry/parser"
 import { resolveRegistryItemsFromRegistries } from "@/src/registry/resolver"
-import {
-  deduplicateFilesByTarget,
-  isLocalFile,
-  isUrl,
-} from "@/src/registry/utils"
-import { Config, getTargetStyleFromConfig } from "@/src/utils/get-config"
+import { isLocalFile, isUrl } from "@/src/registry/utils"
+import { Config } from "@/src/utils/get-config"
 import { getProjectTailwindVersionFromConfig } from "@/src/utils/get-project-info"
 import { handleError } from "@/src/utils/handle-error"
 import { logger } from "@/src/utils/logger"
-import { buildTailwindThemeColorsFromCssVars } from "@/src/utils/updaters/update-tailwind-config"
-import deepmerge from "deepmerge"
 import { z } from "zod"
 
 import { BASE_COLORS } from "./constants"
@@ -29,7 +19,6 @@ import {
   registryBaseColorSchema,
   registryIndexSchema,
   registryItemSchema,
-  registryResolvedItemsTreeSchema,
   registrySchema,
   stylesSchema,
 } from "./schema"
@@ -310,543 +299,6 @@ export async function getRegistryItems(
 
   return results
 }
-
-async function resolveDependenciesRecursively(
-  dependencies: string[],
-  config?: Pick<Config, "style" | "registries" | "resolvedPaths">,
-  visited: Set<string> = new Set()
-) {
-  const items: z.infer<typeof registryItemSchema>[] = []
-  const registryNames: string[] = []
-
-  for (const dep of dependencies) {
-    if (visited.has(dep)) {
-      continue
-    }
-    visited.add(dep)
-
-    // Handle URLs and local files directly
-    if (isUrl(dep) || isLocalFile(dep)) {
-      const item = await getRegistryItem(dep, config)
-      if (item) {
-        items.push(item)
-        if (item.registryDependencies) {
-          const nested = await resolveDependenciesRecursively(
-            item.registryDependencies,
-            config,
-            visited
-          )
-          items.push(...nested.items)
-          registryNames.push(...nested.registryNames)
-        }
-      }
-    }
-    // Handle namespaced items (e.g., @one/foo, @two/bar)
-    else if (dep.startsWith("@") && config?.registries) {
-      // Check if the registry exists
-      const { registry } = parseRegistryAndItemFromString(dep)
-      if (registry && !(registry in config.registries)) {
-        throw new Error(
-          `The items you're adding depend on unknown registry ${registry}. \nMake sure it is defined in components.json as follows:\n` +
-            `{\n  "registries": {\n    "${registry}": "https://example.com/{name}.json"\n  }\n}`
-        )
-      }
-
-      // Let getRegistryItem handle the namespaced item with config
-      // This ensures proper authentication headers are used
-      const item = await getRegistryItem(dep, config)
-      if (item) {
-        items.push(item)
-        if (item.registryDependencies) {
-          const nested = await resolveDependenciesRecursively(
-            item.registryDependencies,
-            config,
-            visited
-          )
-          items.push(...nested.items)
-          registryNames.push(...nested.registryNames)
-        }
-      }
-    }
-    // Handle regular component names
-    else {
-      registryNames.push(dep)
-
-      if (config) {
-        try {
-          const item = await getRegistryItem(dep, config)
-          if (item && item.registryDependencies) {
-            const nested = await resolveDependenciesRecursively(
-              item.registryDependencies,
-              config,
-              visited
-            )
-            items.push(...nested.items)
-            registryNames.push(...nested.registryNames)
-          }
-        } catch (error) {
-          // If we can't fetch the registry item, that's okay - we'll still include the name.
-        }
-      }
-    }
-  }
-
-  return { items, registryNames }
-}
-
-export async function registryResolveItemsTree(
-  names: z.infer<typeof registryItemSchema>["name"][],
-  config: Config
-) {
-  try {
-    // Check for namespaced items when no registries are configured
-    const namespacedItems = names.filter(
-      (name) => !isLocalFile(name) && !isUrl(name) && name.startsWith("@")
-    )
-
-    if (namespacedItems.length > 0 && !config?.registries) {
-      const { registry } = parseRegistryAndItemFromString(namespacedItems[0])
-      throw new Error(
-        `Unknown registry "${registry}". Make sure it is defined in components.json as follows:\n` +
-          `{\n  "registries": {\n    "${registry}": "https://example.com/{name}.json"\n  }\n}`
-      )
-    }
-
-    // Separate local files, URLs, and registry names.
-    const localFiles = names.filter((name) => isLocalFile(name))
-    const urls = names.filter((name) => isUrl(name))
-    const registryNames = names.filter(
-      (name) => !isLocalFile(name) && !isUrl(name)
-    )
-
-    let payload: z.infer<typeof registryItemWithSourceSchema>[] = []
-
-    // Handle local files and URLs directly, resolving their dependencies individually.
-    let allDependencyItems: z.infer<typeof registryItemWithSourceSchema>[] = []
-    let allDependencyRegistryNames: string[] = []
-
-    const resolvedStyle = await getResolvedStyle(config)
-    const configWithStyle =
-      config && resolvedStyle ? { ...config, style: resolvedStyle } : config
-
-    // Deduplicate exact URLs/paths before fetching
-    const uniqueLocalFiles = Array.from(new Set(localFiles))
-    const uniqueUrls = Array.from(new Set(urls))
-
-    for (const localFile of uniqueLocalFiles) {
-      const item = await getRegistryItem(localFile)
-      if (item) {
-        // Add source tracking
-        const itemWithSource: z.infer<typeof registryItemWithSourceSchema> = {
-          ...item,
-          _source: localFile,
-        }
-        payload.push(itemWithSource)
-        if (item.registryDependencies) {
-          // Resolve namespace syntax and set headers for dependencies
-          let resolvedDependencies = item.registryDependencies
-
-          // Check for namespaced dependencies when no registries are configured
-          if (!config?.registries) {
-            const namespacedDeps = item.registryDependencies.filter((dep) =>
-              dep.startsWith("@")
-            )
-            if (namespacedDeps.length > 0) {
-              const { registry } = parseRegistryAndItemFromString(
-                namespacedDeps[0]
-              )
-              throw new Error(
-                `The items you're adding depend on unknown registry ${registry}. \nMake sure it is defined in components.json as follows:\n` +
-                  `{\n  "registries": {\n    "${registry}": "https://example.com/{name}.json"\n  }\n}`
-              )
-            }
-          } else {
-            resolvedDependencies = resolveRegistryItemsFromRegistries(
-              item.registryDependencies,
-              configWithStyle
-            )
-          }
-
-          const { items, registryNames } = await resolveDependenciesRecursively(
-            resolvedDependencies,
-            config,
-            new Set()
-          )
-          allDependencyItems.push(...items)
-          allDependencyRegistryNames.push(...registryNames)
-        }
-      }
-    }
-
-    for (const url of uniqueUrls) {
-      const item = await getRegistryItem(url, config)
-      if (item) {
-        // Add source tracking
-        const itemWithSource: z.infer<typeof registryItemWithSourceSchema> = {
-          ...item,
-          _source: url,
-        }
-        payload.push(itemWithSource)
-        if (item.registryDependencies) {
-          // Resolve namespace syntax and set headers for dependencies
-          let resolvedDependencies = item.registryDependencies
-
-          // Check for namespaced dependencies when no registries are configured
-          if (!config?.registries) {
-            const namespacedDeps = item.registryDependencies.filter((dep) =>
-              dep.startsWith("@")
-            )
-            if (namespacedDeps.length > 0) {
-              const { registry } = parseRegistryAndItemFromString(
-                namespacedDeps[0]
-              )
-              throw new Error(
-                `The items you're adding depend on unknown registry ${registry}. \nMake sure it is defined in components.json as follows:\n` +
-                  `{\n  "registries": {\n    "${registry}": "https://example.com/{name}.json"\n  }\n}`
-              )
-            }
-          } else {
-            resolvedDependencies = resolveRegistryItemsFromRegistries(
-              item.registryDependencies,
-              configWithStyle
-            )
-          }
-
-          const { items, registryNames } = await resolveDependenciesRecursively(
-            resolvedDependencies,
-            config,
-            new Set()
-          )
-
-          allDependencyItems.push(...items)
-          allDependencyRegistryNames.push(...registryNames)
-        }
-      }
-    }
-
-    payload.push(...allDependencyItems)
-
-    // Handle registry names using the new fetchFromRegistry logic.
-    const allRegistryNames = [...registryNames, ...allDependencyRegistryNames]
-    if (allRegistryNames.length > 0) {
-      // Separate namespaced and non-namespaced items
-      const nonNamespacedItems = allRegistryNames.filter(
-        (name) => !name.startsWith("@")
-      )
-      const namespacedItems = allRegistryNames.filter((name) =>
-        name.startsWith("@")
-      )
-
-      // Handle namespaced items directly with getRegistryItems
-      if (namespacedItems.length > 0) {
-        const results = await getRegistryItems(namespacedItems, config)
-
-        // Filter out null results and track successful items
-        const successfulResults: Array<{
-          item: z.infer<typeof registryItemSchema>
-          index: number
-        }> = []
-        results.forEach((result, index) => {
-          if (result) {
-            successfulResults.push({ item: result, index })
-          } else {
-            logger.warn(
-              `Failed to fetch namespaced item: ${namespacedItems[index]}`
-            )
-          }
-        })
-
-        // Add source tracking for successful namespaced items
-        const itemsWithSource: z.infer<typeof registryItemWithSourceSchema>[] =
-          successfulResults.map(({ item, index }) => ({
-            ...item,
-            _source: namespacedItems[index],
-          }))
-
-        payload.push(...itemsWithSource)
-
-        // Process dependencies of namespaced items
-        for (const { item } of successfulResults) {
-          if (item.registryDependencies) {
-            const { items: depItems, registryNames: depNames } =
-              await resolveDependenciesRecursively(
-                item.registryDependencies,
-                config,
-                new Set([...namespacedItems])
-              )
-            payload.push(...depItems)
-
-            // Add any non-namespaced dependencies to be processed
-            const nonNamespacedDeps = depNames.filter(
-              (name) => !name.startsWith("@")
-            )
-            nonNamespacedItems.push(...nonNamespacedDeps)
-          }
-        }
-      }
-
-      // For non-namespaced items, we need the index and style resolution
-      if (nonNamespacedItems.length > 0) {
-        const index = await getRegistryIndex()
-        if (!index && payload.length === 0) {
-          return null
-        }
-
-        if (index) {
-          // Remove duplicates from non-namespaced items
-          const uniqueNonNamespaced = Array.from(new Set(nonNamespacedItems))
-
-          // If we're resolving the index, we want it to go first
-          if (uniqueNonNamespaced.includes("index")) {
-            uniqueNonNamespaced.unshift("index")
-          }
-
-          // Resolve non-namespaced items through the existing flow
-          let registryItems = await resolveRegistryItems(
-            uniqueNonNamespaced,
-            config
-          )
-          let result = await fetchRegistry(registryItems)
-          const registryPayload = z.array(registryItemSchema).parse(result)
-          payload.push(...registryPayload)
-        }
-      }
-    }
-
-    if (!payload.length) {
-      return null
-    }
-
-    // No deduplication - we want to support multiple items with the same name from different sources
-
-    // If we're resolving the index, we want to fetch
-    // the theme item if a base color is provided.
-    // We do this for index only.
-    // Other components will ship with their theme tokens.
-    if (allRegistryNames.includes("index")) {
-      if (config.tailwind.baseColor) {
-        const theme = await registryGetTheme(config.tailwind.baseColor, config)
-        if (theme) {
-          payload.unshift(theme)
-        }
-      }
-    }
-
-    // Build source map for topological sort
-    const sourceMap = new Map<z.infer<typeof registryItemSchema>, string>()
-    payload.forEach((item) => {
-      // Use the _source property if it was added, otherwise use the name
-      const source = item._source || item.name
-      sourceMap.set(item, source)
-    })
-
-    // Apply topological sort to ensure dependencies come before dependents
-    payload = topologicalSortRegistryItems(payload, sourceMap)
-
-    // Sort the payload so that registry:theme items come first,
-    // while maintaining the relative order of all items.
-    payload.sort((a, b) => {
-      if (a.type === "registry:theme" && b.type !== "registry:theme") {
-        return -1
-      }
-      if (a.type !== "registry:theme" && b.type === "registry:theme") {
-        return 1
-      }
-      return 0
-    })
-
-    let tailwind = {}
-    payload.forEach((item) => {
-      tailwind = deepmerge(tailwind, item.tailwind ?? {})
-    })
-
-    let cssVars = {}
-    payload.forEach((item) => {
-      cssVars = deepmerge(cssVars, item.cssVars ?? {})
-    })
-
-    let css = {}
-    payload.forEach((item) => {
-      css = deepmerge(css, item.css ?? {})
-    })
-
-    let docs = ""
-    payload.forEach((item) => {
-      if (item.docs) {
-        docs += `${item.docs}\n`
-      }
-    })
-
-    let envVars = {}
-    payload.forEach((item) => {
-      envVars = deepmerge(envVars, item.envVars ?? {})
-    })
-
-    // Deduplicate files based on resolved target paths.
-    const deduplicatedFiles = await deduplicateFilesByTarget(
-      payload.map((item) => item.files ?? []),
-      config
-    )
-
-    const parsed = registryResolvedItemsTreeSchema.parse({
-      dependencies: deepmerge.all(
-        payload.map((item) => item.dependencies ?? [])
-      ),
-      devDependencies: deepmerge.all(
-        payload.map((item) => item.devDependencies ?? [])
-      ),
-      files: deduplicatedFiles,
-      tailwind,
-      cssVars,
-      css,
-      docs,
-    })
-
-    if (Object.keys(envVars).length > 0) {
-      parsed.envVars = envVars
-    }
-
-    return parsed
-  } catch (error) {
-    handleError(error)
-    return null
-  }
-}
-
-async function resolveRegistryDependencies(url: string, config: Config) {
-  if (isUrl(url)) {
-    return [url]
-  }
-
-  const { registryNames } = await resolveDependenciesRecursively(
-    [url],
-    config,
-    new Set()
-  )
-
-  const style = config.resolvedPaths?.cwd
-    ? await getTargetStyleFromConfig(config.resolvedPaths.cwd, config.style)
-    : config.style
-
-  const urls = registryNames.map((name) =>
-    resolveRegistryUrl(isUrl(name) ? name : `styles/${style}/${name}.json`)
-  )
-
-  return Array.from(new Set(urls))
-}
-
-export async function registryGetTheme(name: string, config: Config) {
-  const [baseColor, tailwindVersion] = await Promise.all([
-    getRegistryBaseColor(name),
-    getProjectTailwindVersionFromConfig(config),
-  ])
-  if (!baseColor) {
-    return null
-  }
-
-  // TODO: Move this to the registry i.e registry:theme.
-  const theme = {
-    name,
-    type: "registry:theme",
-    tailwind: {
-      config: {
-        theme: {
-          extend: {
-            borderRadius: {
-              lg: "var(--radius)",
-              md: "calc(var(--radius) - 2px)",
-              sm: "calc(var(--radius) - 4px)",
-            },
-            colors: {},
-          },
-        },
-      },
-    },
-    cssVars: {
-      theme: {},
-      light: {
-        radius: "0.5rem",
-      },
-      dark: {},
-    },
-  } satisfies z.infer<typeof registryItemSchema>
-
-  if (config.tailwind.cssVariables) {
-    theme.tailwind.config.theme.extend.colors = {
-      ...theme.tailwind.config.theme.extend.colors,
-      ...buildTailwindThemeColorsFromCssVars(baseColor.cssVars.dark ?? {}),
-    }
-    theme.cssVars = {
-      theme: {
-        ...baseColor.cssVars.theme,
-        ...theme.cssVars.theme,
-      },
-      light: {
-        ...baseColor.cssVars.light,
-        ...theme.cssVars.light,
-      },
-      dark: {
-        ...baseColor.cssVars.dark,
-        ...theme.cssVars.dark,
-      },
-    }
-
-    if (tailwindVersion === "v4" && baseColor.cssVarsV4) {
-      theme.cssVars = {
-        theme: {
-          ...baseColor.cssVarsV4.theme,
-          ...theme.cssVars.theme,
-        },
-        light: {
-          radius: "0.625rem",
-          ...baseColor.cssVarsV4.light,
-        },
-        dark: {
-          ...baseColor.cssVarsV4.dark,
-        },
-      }
-    }
-  }
-
-  return theme
-}
-
-// TODO: We're double-fetching here. Use a cache.
-export async function resolveRegistryItems(names: string[], config: Config) {
-  let registryDependencies: string[] = []
-
-  const registryNames = names.filter(
-    (name) => !isLocalFile(name) && !isUrl(name)
-  )
-
-  const resolvedStyle = await getResolvedStyle(config)
-  for (const name of registryNames) {
-    let resolvedName = name
-    if (config) {
-      try {
-        const configWithStyle =
-          config && resolvedStyle ? { ...config, style: resolvedStyle } : config
-
-        const resolved = buildUrlAndHeadersForRegistryItem(
-          name,
-          configWithStyle
-        )
-        if (resolved) {
-          resolvedName = resolved.url
-        }
-      } catch (error) {}
-    }
-
-    const itemRegistryDependencies = await resolveRegistryDependencies(
-      resolvedName,
-      config
-    )
-    registryDependencies.push(...itemRegistryDependencies)
-  }
-
-  return Array.from(new Set(registryDependencies))
-}
-
 export function getRegistryTypeAliasMap() {
   return new Map<string, string>([
     ["registry:ui", "ui"],
